:recycle: refactor: Offset 사용하도록 변경
